Wrap all event points in div to fix highlighting
Previously, on Chrome, row and column highlight on hover would highlight
the table cell as a circle instead of treating it independently from
row, leading weird white spaces in the corners of the cell. (Note to
self: you took a screenshot of this behavior, maybe post it in an issue
or something.)

This increases the number of DOM elements, so ideally this would not be
needed. Might be possible to revert if Chrome ever fixes the issue.