Refactor implement remaining checkbox logic
Decided not to be lazy. Yes I checked/unchecked 60 checkboxes to test
this...

On another note, seems to be odd bug in the layout where selecting or
deselecting all teams causes the table to shift, but the same effect
doesn't occur when going from one team to no teams and vice versa. Only
noticable on a wide desktop layout when there is space to either side of
the table. Might be because of the scrollbar on Firefox? If so, easy
fix.