Fix bug where medal highlights weren't updating
When the event focused column is added by selecting an event to sort by,
the medal highlights would remain even if turned off earlier. This is
now fixed.

This JS is a mess, I should really learn how to write good JS, if such a
thing exists. Surely it does?