Add 2016 Nats results
Did not include the exhibition team in Div C for now (because the
conversion script couldn't handle it properly), but will try to add
their results for all the Nats years later (will write a script to
modifiy an existing SciolyFF file by shifting down all the placings).

Not going to bother to document the conversion process because hopefully
no one will ever have to do it again, but for the most part just
involved a lot of Excel copy-pasting and text-to-columns and some string
splitting formulas.