Add scroll to top button and clean up anchors
Prevents anchors from being added to URL so that back button behavior is
more consistent, but have it degrade gracefully when JS is disabled.

Had an issue where in development, Webpack doesn't eliminate duplication
so some CSS was overriding things it shouldn't have, thus requiring the
import of button-float in both results_index and results_template. Might
add FABs in results_template in the future so this isn't a huge deal.