Use `position: sticky` only on th instead of thead
Fixes issue in iOS 13 browsers where sticky is applied twice (despite
one of the rules not being prefixed with webkit). Degrades performance
on browsers that support sticky thead but limit the number of elements
that can be "insta-stickied", like Firefox for Android. May revert if
another solution can be found for iOS 13.