Add WA NW Regionals (B & C), courtesy Tailsfan101
As pointed out by Tailsfan101, there are some discrepencies with scoring
(for Div. C, the no-show points seem to be off and for Div B. some of
the team scores are off by one or so).

Ignoring these for now until we can deterine the root cause, so the
scores on Unosmium will be different than the original official results.